+++ examples/with_file.js
@@ -14,7 +14,7 @@
  */
 // import
 try{
-    var monitor = require('../index.js'); // use 'monitode' instead
+    var monitode = require('../index.js'); // use 'monitode' instead
 } catch (MODULE_NOT_FOUND){
     console.log(MODULE_NOT_FOUND);
     process.exit(1);
@@ -22,7 +22,7 @@ try{
 
 // using standalone
 var file = __dirname + '/example.txt';
-monitor({
+monitode({
     output: true,
     logger: {
         file: file,