+++ examples/with_mail.js
@@ -14,14 +14,14 @@
  */
 // import
 try{
-    var monitor = require('../index.js'); // use 'monitode' instead
+    var monitode = require('../index.js'); // use 'monitode' instead
 } catch (MODULE_NOT_FOUND){
     console.log(MODULE_NOT_FOUND);
     process.exit(1);
 }
 
 // using standalone
-monitor({
+monitode({
     mail: {
         provider: 'gmail',
         user: 'pippo@gmail.com',